+++ pkg/ansible/run.go
@@ -173,11 +173,11 @@ func Run(flags *aoflags.AnsibleOperatorFlags) error {
 	services := []*v1.Service{service}
 	_, err = metrics.CreateServiceMonitors(cfg, namespace, services)
 	if err == metrics.ErrServiceMonitorNotPresent {
-		log.Info("Could not create ServiceMonitor object", "reason", err.Error())
+		log.Info("Could not create ServiceMonitor object", "error", err.Error())
 		// If this operator is deployed to a cluster without the prometheus-operator running, it will return
 		// ErrServiceMonitorNotPresent, which can be used to safely skip ServiceMonitor creation.
 		if err == metrics.ErrServiceMonitorNotPresent {
-			log.Info("Install prometheus-operator in your cluster to create ServiceMonitor objects", "reason", err.Error())
+			log.Info("Install prometheus-operator in your cluster to create ServiceMonitor objects", "error", err.Error())
 		}
 	}